src/main/java/com/scott/betaexam/service/EmployeeServiceImpl.java
package com.scott.betaexam.service;

import com.scott.betaexam.model.Employee;
import com.scott.betaexam.repo.EmployeeRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class EmployeeServiceImpl implements EmployeeService {

    @Autowired
    private EmployeeRepository employeeRepository;

    public List<Employee> getEmployees(String name) {
        return employeeRepository.findByNameContaining(name);
    }


    public Employee getEmployeeByEmail(String email) {
        return employeeRepository.findByEmail(email);
    }

    @Override
    public Employee saveEmployee(Employee employee) {
        return employeeRepository.save(employee);
    }
}
